Cleaned up most of the CMakeLists.txt files for the project Moved some of the packets into their own files.
