+++ src/CMakeLists.txt
@@ -2,8 +2,8 @@ CMAKE_MINIMUM_REQUIRED(VERSION 2.8 FATAL_ERROR)
 PROJECT(src)
 
 IF(ENABLE_TESTING_COVERAGE)
-    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -O0 -fprofile-arcs -ftest-coverage")
-    SET(CMAKE_C_FLAGS "${CMAKE_CXX_FLAGS} -g -O0 -fprofile-arcs -ftest-coverage")
+  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -O0 -fprofile-arcs -ftest-coverage")
+  SET(CMAKE_C_FLAGS "${CMAKE_CXX_FLAGS} -g -O0 -fprofile-arcs -ftest-coverage")
 ENDIF() # ENABLE_TESTING_COVERAGE
 
 IF(ENABLE_TESTING)
@@ -16,12 +16,12 @@ IF(ENABLE_TESTING)
 ENDIF() # ENABLE_TESTING
 
 IF( ENABLE_ASAN )
-	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -fsanitize=address -fno-omit-frame-pointer")
-	SET(LINK_FLAGS "${LINK_FLAGS} -g -fsanitize=address")
+  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -fsanitize=address -fno-omit-frame-pointer")
+  SET(LINK_FLAGS "${LINK_FLAGS} -g -fsanitize=address")
 ENDIF()
 
 IF( DISABLE_CRYPT )
-    SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DDISABLE_CRYPT")
+  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DDISABLE_CRYPT")
 ENDIF()
 
 INCLUDE_DIRECTORIES(
@@ -30,25 +30,25 @@ INCLUDE_DIRECTORIES(
   SYSTEM ${EXTERNAL_INCLUDE_DIR}
   SYSTEM ${PBJSON_ROOT}
   SYSTEM ${SPDLOG_ROOT}
-	SYSTEM ${SELENE_ROOT}
-	SYSTEM ${SOL2_ROOT}
-	SYSTEM ${LUA_INCLUDE_DIR}
-	SYSTEM ${CXXOPT_ROOT}
+  SYSTEM ${SELENE_ROOT}
+  SYSTEM ${SOL2_ROOT}
+  SYSTEM ${LUA_INCLUDE_DIR}
+  SYSTEM ${CXXOPT_ROOT}
 )
 
 LINK_DIRECTORIES(
-	${EXTERNAL_BINARY_DIR}
+  ${EXTERNAL_BINARY_DIR}
 )
 
 IF(ENABLE_MYSQL)
   INCLUDE_DIRECTORIES(
-      SYSTEM ${MYSQL_INCLUDE_DIRS}
-      SYSTEM ${MYSQLpp_INCLUDE_DIR}
+    SYSTEM ${MYSQL_INCLUDE_DIRS}
+    SYSTEM ${MYSQLpp_INCLUDE_DIR}
   )
   
   LINK_DIRECTORIES(
-	  ${MYSQL_LIBRARY_PATH}
-	  ${MYSQLpp_LIBRARY_PATH}
+    ${MYSQL_LIBRARY_PATH}
+    ${MYSQLpp_LIBRARY_PATH}
   )
   
   IF (MSVC)
@@ -69,6 +69,6 @@ ADD_SUBDIRECTORY(char)
 ADD_SUBDIRECTORY(map)
 
 IF(ENABLE_TESTING)
-    ADD_SUBDIRECTORY(tests)
+  ADD_SUBDIRECTORY(tests)
 ENDIF()
