+++ src/char/CMakeLists.txt
@@ -30,6 +30,7 @@ ENDIF()
 INCLUDE_DIRECTORIES(
     ${CMAKE_SOURCE_DIR}/src/core/include
     ${CMAKE_SOURCE_DIR}/src/rosecommon/include
+	${CMAKE_SOURCE_DIR}/src/rosecommon/include/packets
     ${CMAKE_SOURCE_DIR}/src/proto/include
     ${CMAKE_CURRENT_SOURCE_DIR}/include
 )
