+++ src/map/CMakeLists.txt
@@ -30,6 +30,7 @@ include_directories(
 	${CMAKE_SOURCE_DIR}/src/core/include
 	${CMAKE_SOURCE_DIR}/src/proto/include
 	${CMAKE_SOURCE_DIR}/src/rosecommon/include
+	${CMAKE_SOURCE_DIR}/src/rosecommon/include/packets
 	${CMAKE_CURRENT_SOURCE_DIR}/include
 )
