+++ src/proto/CMakeLists.txt
@@ -4,13 +4,6 @@ SET(PROJECT protoGen)
 project(${PROJECT})
 
 IF (MSVC)
-#foreach(flag_var
-#        CMAKE_CXX_FLAGS CMAKE_CXX_FLAGS_DEBUG CMAKE_CXX_FLAGS_RELEASE
-#        CMAKE_CXX_FLAGS_MINSIZEREL CMAKE_CXX_FLAGS_RELWITHDEBINFO)
-#   if(${flag_var} MATCHES "/MD")
-#      string(REGEX REPLACE "/MD" "/MT" ${flag_var} "${${flag_var}}")
-#   endif(${flag_var} MATCHES "/MD")
-#endforeach(flag_var)
   add_definitions(-D_SCL_SECURE_NO_WARNINGS)
 ELSE()
 SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
@@ -25,7 +18,6 @@ ENDIF()
 include_directories(
 	${PROTOBUF_INCLUDE_DIRS}
 )
-
 
 SET( ProtoFiles
 	configfile.proto
