+++ src/rosecommon/CMakeLists.txt
@@ -25,7 +25,8 @@ ENDIF()
 INCLUDE_DIRECTORIES(
 	${CMAKE_SOURCE_DIR}/src/core/include
 	${CMAKE_SOURCE_DIR}/src/proto/include
-        ${CMAKE_CURRENT_SOURCE_DIR}/include
+	${CMAKE_CURRENT_SOURCE_DIR}/include
+	${CMAKE_CURRENT_SOURCE_DIR}/include/packets
 )
 
 SET( SOURCES_HEADER
