+++ src/rosecommon/include/mappackets.h
@@ -17,260 +17,27 @@
 
 #include "epackettype.h"
 #include "crosepacket.h"
+#include "dataconsts.h"
 #include <string>
 #include <exception>
 #include <vector>
 #include <array>
 
+
+#include "cli_chat.h"
+#include "cli_stop.h"
+#include "cli_logout.h"
+#include "cli_revive.h"
+#include "cli_changemap.h"
+#include "srv_chat.h"
+#include "src_revive.h"
+#include "srv_logout.h"
+#include "srv_initdata.h"
+#include "srv_changemap.h"
+#include "srv_serverdata.h"
+#include "srv_removeobject.h"
+
 namespace RoseCommon {
-#define MIN_SELL_TYPE 1
-#define MAX_SELL_TYPE 11
-
-//------------------------------------------------
-//------------------------------------------------
-
-class CliLogoutReq : public CRosePacket {
- public:
-  CliLogoutReq(uint8_t buffer[MAX_PACKET_SIZE]);
-  CliLogoutReq();
-
-  virtual ~CliLogoutReq();
-};
-
-//------------------------------------------------
-
-class SrvLogoutReply : public CRosePacket {
- public:
-  SrvLogoutReply(uint16_t wait_time);
-
-  virtual ~SrvLogoutReply();
-
-  uint16_t wait_time() const;
-
- protected:
-  void pack();
-
- private:
-  uint16_t wait_time_;
-};
-
-//------------------------------------------------
-//------------------------------------------------
-
-class CliChangeMapReq : public CRosePacket {
- public:
-  CliChangeMapReq(uint8_t buffer[MAX_PACKET_SIZE]);
-  CliChangeMapReq();
-
-  virtual ~CliChangeMapReq();
-
-  uint8_t weight_rate() const;
-  uint16_t position_z() const;
-
- protected:
-  void pack();
-
- private:
-  uint8_t weight_rate_;
-  uint16_t position_z_;  // this is not actually set
-};
-
-//------------------------------------------------
-
-class SrvChangeMapReply : public CRosePacket {
- public:
-  SrvChangeMapReply(uint16_t object_index, uint16_t current_hp,
-                    uint16_t current_mp, uint16_t current_exp,
-                    uint16_t penalize_exp, uint16_t world_time,
-                    uint16_t team_number);
-
-  virtual ~SrvChangeMapReply();
-
-  uint16_t object_index() const;
-  uint16_t current_hp() const;
-  uint16_t current_mp() const;
-
-  void setItemRate(uint8_t type, uint8_t rate);
-
- protected:
-  void pack();
-
- private:
-  struct global_var {
-    uint16_t craft_rate_;
-    uint32_t update_time_;
-    uint16_t world_rate_;
-    uint8_t town_rate_;
-    uint8_t item_rate_[MAX_SELL_TYPE];
-    uint32_t flags_;
-
-    global_var() : craft_rate_(0),
-    update_time_(0),
-    world_rate_(0),
-    town_rate_(0),
-    flags_(0) {
-      for (int i = 0; i < MAX_SELL_TYPE; ++i) {
-        item_rate_[i] = 0;
-      }
-    }
-  };
-
-  uint16_t object_index_;
-  uint16_t current_hp_;
-  uint16_t current_mp_;
-  uint64_t current_exp_;
-  uint64_t penalize_exp_;
-  uint32_t world_time_;
-  uint32_t team_number_;
-  global_var zone_vars_;
-};
-
-//------------------------------------------------
-//------------------------------------------------
-
-class CliChat : public CRosePacket {
- public:
-  CliChat(uint8_t buffer[MAX_PACKET_SIZE]);
-  CliChat(const std::string &chat = "");
-
-  std::string getMessage() const;
-
- protected:
-  void pack();
-
- private:
-  std::string chat_;
-};
-
-//------------------------------------------------
-
-class SrvChat : public CRosePacket {
- public:
-  SrvChat(const std::string &chat = "", uint16_t charuid = 0)
-      : CRosePacket(ePacketType::PAKWC_NORMAL_CHAT),
-        chat_(chat),
-        charuid_(charuid) {}
-
- protected:
-  void pack() { *this << charuid_ << chat_; }
-
- private:
-  std::string chat_;
-  uint16_t charuid_;
-};
-
-//------------------------------------------------
-//------------------------------------------------
-
-class CliReviveReq : public CRosePacket {
- public:
-  CliReviveReq(uint8_t buffer[MAX_PACKET_SIZE]);
-  CliReviveReq(uint8_t type = 0);
-
-  enum eType : uint8_t { REVIVE_POS, SAVE_POS, START_POS, CURRENT_POS };
-
-  eType getType() const;
-
- protected:
-  void pack();
-
- private:
-  uint8_t type_;
-};
-
-//------------------------------------------------
-
-class SrvReviveReply : public CRosePacket {
- public:
-  SrvReviveReply(uint16_t mapid = 0);
-
- protected:
-  void pack();
-
- private:
-  uint16_t mapid_;
-};
-
-//------------------------------------------------
-//------------------------------------------------
-
-class SrvInitDataReply : public CRosePacket {
- public:
-  SrvInitDataReply(uint32_t rand_seed, uint16_t rand_index);
-
-  virtual ~SrvInitDataReply();
-
-  uint32_t rand_seed() const;
-  uint16_t rand_index() const;
-
- protected:
-  void pack();
-
- private:
-  uint32_t rand_seed_;
-  uint16_t rand_index_;
-};
-
-//------------------------------------------------
-//------------------------------------------------
-
-class SrvServerData : public CRosePacket {
- public:
-  SrvServerData(uint8_t type);
-
-  virtual ~SrvServerData();
-
-  uint8_t type() const;
-
-  enum data_type : uint8_t { ECONOMY = 0, NPC };
-
- protected:
-  void pack();
-
- private:
-  struct Enconmy_Data {
-    uint32_t counter_;
-    uint16_t pop_base_;
-    uint16_t dev_base_;
-    uint16_t consume_[MAX_SELL_TYPE];
-    uint16_t dev_;
-    uint32_t pop_;
-    uint32_t item_[MAX_SELL_TYPE];
-
-    Enconmy_Data() : counter_(0),
-    pop_base_(0),
-    dev_base_(0),
-    dev_(0),
-    pop_(0) {
-      for(int i =0; i < MAX_SELL_TYPE; ++i) {
-        item_[i] = consume_[i] = 0;
-      }
-    }
-  };
-  uint8_t type_;
-  Enconmy_Data enconmy_data_;
-};
-
-//------------------------------------------------
-//------------------------------------------------
-
-class SrvRemoveObject : public CRosePacket {
- public:
-  SrvRemoveObject(uint16_t obj_id);
-
- protected:
-  void pack();
-
- private:
-  uint16_t obj_id_;
-};
-
-//------------------------------------------------
-//------------------------------------------------
-#define MAX_UNION_COUNT 10
-#define MAX_BUFF_STATUS 40
-#define MAX_SKILL_COUNT 120
-#define MAX_HOTBAR_ITEMS 32
 
 class SrvSelectCharReply : public CRosePacket {
  public:
@@ -588,17 +355,7 @@ class SrvBillingMsg : public CRosePacket {
   void pack() { *this << (uint16_t)0x1001 << (uint32_t)2; };
 };
 
-class CliStopReq : public CRosePacket {
- public:
-  CliStopReq(uint8_t buffer[MAX_PACKET_SIZE]);
-  virtual ~CliStopReq() {};
 
-  float getX() const { return x; }
-  float getY() const { return y; }
-
-private:
-  float x, y;
-};
 
 //------------------------------------------------
 //------------------------------------------------
